backend/requirements.txt
Django==5.0.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
firebase-admin==6.6.0
djangorestframework==3.15.2
django-cors-headers==4.6.0